requirement.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.2
djangorestframework==3.12.4
idna==2.10
pytube==10.7.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
